requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-google-genai
PyMuPDF
docx2txt
faiss-cpu